Refactor home screen
